Timerec parse performance enhancement.
